pyproject.toml
[project]
name = "repl-nix-workspace"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "email-validator>=2.2.0",
    "flask-apscheduler>=1.13.1",
    "flask>=3.1.0",
    "flask-sqlalchemy>=3.1.1",
    "gunicorn>=23.0.0",
    "psycopg2-binary>=2.9.10",
    "telegram>=0.0.1",
    "pandas>=2.2.3",
    "python-telegram-bot>=21.11.1",
    "pyjwt>=2.10.1",
    "openpyxl>=3.1.5",
    "numpy>=2.2.3",
    "requests>=2.32.3",
    "sqlalchemy>=2.0.38",
]
